requirements-pc.txt
# PC requirements for PiEdge EduKit
# Install these for training and development on PC

# Core ML frameworks
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0
onnx>=1.14.0,<1.16.0
onnxruntime>=1.15.0,<1.17.0
onnxruntime-tools>=1.15.0,<1.17.0

# Data processing
numpy>=1.24.0,<1.27.0
Pillow>=9.5.0,<11.0.0
pandas>=2.0.0,<2.2.0
scikit-learn>=1.3.0,<1.5.0

# Visualization and analysis
matplotlib>=3.7.0,<3.9.0
streamlit>=1.28.0,<1.30.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
tqdm>=4.65.0,<4.67.0

# System monitoring
psutil>=5.9.0,<6.0.0

